Move token budget classes to support packages
