+++ embabel-agent-common/embabel-agent-ai/src/main/kotlin/com/embabel/common/ai/model/TokenCounter.kt
@@ -13,7 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package com.embabel.common.ai.model
 
 import org.jetbrains.annotations.ApiStatus
